추천게시글 dto swagger 적용
