--- back-end/src/main/java/com/dreamteam/hola/dto/RecommendedBoardDto.java
@@ -1,13 +1,0 @@
-package com.dreamteam.hola.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class RecommendedBoardDto {
-    private Long  boardId;
-    private String title;
-}
